Now use kf * A * B - kr * C formalism instead of kr * Kc * A * B - kr * C for RNAP binding to a promoter alone.
Changed species compartment assignments to "default" to avoid clashing while we figure out what we're doing with compartments.

Revised decay abstraction for sequestered species S1 so that if kd of S1 is equal to all those of its sequestering complexes then the decay rate law will simply be kd * S1.