Abstracted equilibrium expressions for complexes now incorporate non-zero degradation.  When the abstracted complex can degrade, then Kc is replaced by kf / (kr + kd).
Notation for forward and reverse rate constants has been changed slightly to reflect the equilibrium binding constants from which they are derived.  For example, kf is now kf_c if it is derived from the eq binding constant Kc.

Also fixed complex visitor bug in which Kc wouldn't be added as a parameter to the abstracted rate law for an output complex when its stoichiometry of binding was greater than 3.