fixed some quirks/bugs related to component moving/snapping.  everything should be centered now.
also, the i reverted the previous change to schematic about the toolbar.  the toolbar is meant to change to a grid-specific one when you put a grid on the schematic.  currently, i haven't disabled having and grid plus other stuff on the schematic.  i'm going to do this soon, so it's either grid or non-grid.