SbolSynthesizer now adds URIs to SequenceAnnotations when saving out SBOL.  (Not yet using URIs for SequenceAnnotations or DnaSequences to full potential.)
Commented out code that generates a hidden implicit promoter when a complex influence is drawn with no promoters present.  This was useful back when we could change the type of a complex influence to activation or repression and needed a promoter to fall back on, but this is no longer the case.  A potentially hidden promoter can also cause problems when saving SBOL since said promoter will be missing SBOL associations.