Modified some SBOLUtility methods for increased efficiency.
Also changed SBOLSynthesizer such that saving a newly synthesized DNA component will result in a check for any previously synthesized component.  If they do not match, then the previously synthesized component is deleted and a new URI is constructed for the new component.

Also fixed bug that allowed multiple placeholder URIs to be added to the SBOLAssociationPanel.